unit tests and fixes for Compile()
