+++ .github/workflows/test.yml
@@ -4,7 +4,7 @@ jobs:
   test:
     strategy:
       matrix:
-        go-version: [1.18.x, 1.19.x]
+        go-version: [1.18.x, 1.19.x, 1.20.x]
         os: [ubuntu-latest]
     runs-on: ${{ matrix.os }}
     steps:
@@ -23,7 +23,7 @@ jobs:
      - name: Install Go
        uses: actions/setup-go@v2
        with: 
-        go-version: 1.19.x
+        go-version: 1.21.x
      - name: Checkout code
        uses: actions/checkout@v2
      - name: Test
