+++ go.mod
@@ -2,9 +2,11 @@ module github.com/hydronica/go-openapi
 
 go 1.19
 
+require github.com/hydronica/trial v0.7.2
+
 require (
-	github.com/davecgh/go-spew v1.1.1
-	github.com/hydronica/trial v0.7.1
+	github.com/google/go-cmp v0.5.9
+	github.com/jbsmith7741/go-tools v0.4.1
 )
 
-require github.com/google/go-cmp v0.5.9 // indirect
+require github.com/pkg/errors v0.9.1 // indirect
